lzma_rs::decompress
Struct Stream

pub struct Stream<W>
where
    W: Write,
{ /* private fields */ }

Lzma decompressor that can process multiple chunks of data using the io::Write interface.

Implementations

impl<W> Stream<W>
where W: Write,

pub fn new(output: W) -> Self

Initialize the stream. This will consume the output which is the sink implementing io::Write that will receive decompressed bytes.

pub fn new_with_options(options: &Options, output: W) -> Self

Initialize the stream with the given options. This will consume the output which is the sink implementing io::Write that will receive decompressed bytes.

pub fn get_output(&self) -> Option<&W>

Get a reference to the output sink

pub fn get_output_mut(&mut self) -> Option<&mut W>

Get a mutable reference to the output sink

pub fn finish(self) -> Result<W>

Consumes the stream and returns the output sink. This also makes sure we have properly reached the end of the stream.